+++ src/Shopway.Domain/Users/Authorization/Permission.Product.cs
