+++ src/Shopway.Presentation/Authentication/RolePermissionAuthentication/IUserAuthorizationService.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Shopway.Domain.Common.Enums;
 using Shopway.Domain.Common.Results;
-using Shopway.Domain.Enums;
 using Shopway.Domain.Users;
+using Shopway.Domain.Users.Authorization;
 
 namespace Shopway.Presentation.Authentication.RolePermissionAuthentication;
