+++ src/Shopway.Presentation/Authentication/RolePermissionAuthentication/RequiredPermissionsAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Shopway.Domain.Common.Enums;
-using Shopway.Domain.Enums;
+using Shopway.Domain.Users.Authorization;
 
 namespace Shopway.Presentation.Authentication.RolePermissionAuthentication;
