+++ test/Shopway.Tests.Integration.Container/Api/TestUserAuthorizationService.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Shopway.Domain.Common.Enums;
 using Shopway.Domain.Common.Results;
-using Shopway.Domain.Enums;
 using Shopway.Domain.Users;
+using Shopway.Domain.Users.Authorization;
 using Shopway.Presentation.Authentication.RolePermissionAuthentication;
 
 namespace Shopway.Tests.Integration.Container.Api;
